NDRS-372: rename AuctionProvider -> Auction, reposition
